pyproject.toml
[project]
name = "streamlined-releases"
version = "0.0.1"
description = "Streamlined Releases for Python Projects"
readme = "README.md"
requires-python = ">=3.12"
classifiers = ["Private :: Do Not Upload"]
dependencies = [
    "git-cliff>=2.9.1",
    "gitpython>=3.1.44",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pygithub>=2.6.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "autopep8>=2.3.2",
    "isort>=6.0.1",
    "sort-all>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = [".vscode"]

[tool.hatch.build.targets.wheel]
packages = ["src/streamlined_releases"]
